+++ apps/resell-api/src/components/member/member.service.ts
@@ -158,28 +158,31 @@ export class MemberService {
 
         if(!result) throw new InternalServerErrorException(Message.SOMETHING_WENT_WRONG);
         
-        const liker = await this.getMember(null, memberId);
-        const notification:NotificationInput= {
-            notificationType: NotificationType.LIKE,
-            notificationGroup: NotificationGroup.MEMBER,
-            notificationTitle: `${liker.memberNick} has liked you`,
-            authorId: memberId,
-            receiverId: likeRefId,
-            productId: null,
-            articleId: null
-        };
-        if(modifier === 1){
-           const result =  await this.notificationService.createNotification(notification);
-           console.log('result ----->',result);
-        }else{
-            const input = {
-                authorId:memberId, 
-                receiverId:likeRefId, 
-                productId:null,
-                articleId:null
+        if(memberId.toString() !== likeRefId.toString()){
+            const liker = await this.getMember(null, memberId);
+            const notification:NotificationInput= {
+                notificationType: NotificationType.LIKE,
+                notificationGroup: NotificationGroup.MEMBER,
+                notificationTitle: `${liker.memberNick} has liked you`,
+                authorId: memberId,
+                receiverId: likeRefId,
+                productId: null,
+                articleId: null
             };
-            await this.notificationService.deleteNotification(input);
+            if(modifier === 1){
+               const result =  await this.notificationService.createNotification(notification);
+               console.log('result ----->',result);
+            }else{
+                const input = {
+                    authorId:memberId, 
+                    receiverId:likeRefId, 
+                    productId:null,
+                    articleId:null
+                };
+                await this.notificationService.deleteNotification(input);
+            }
         }
+        
         return result;
     };
